Moved CardGen to __init__.py
